feat(glossary): method and function
function
A series of statements which returns some value to a caller. It can also be passed zero or more arguments which may be used in the execution of the body.

method
A function which is defined inside a class body. If called as an attribute of an instance of that class, the method will get the instance object as its first argument (which is usually called self).